[add]レイアウトの調整

+++ app/controllers/public/bookmarks_controller.rb
@@ -4,16 +4,25 @@ class Public::BookmarksController < ApplicationController
     @bookmarks = current_user.bookmarks.page(params[:page]).per(10)
   end
   def create
-    sauna = Sauna.find(params[:sauna_id])
-    bookmark = current_user.bookmarks.new(sauna_id: sauna.id)
-    bookmark.save
-    redirect_back(fallback_location: root_path)
+    if user_signed_in?
+      sauna = Sauna.find(params[:sauna_id])
+      bookmark = current_user.bookmarks.new(sauna_id: sauna.id)
+      bookmark.save
+      redirect_back(fallback_location: root_path)
+    else
+      redirect_to new_user_session_path
+    end
   end
 
   def destroy
-    sauna = Sauna.find(params[:sauna_id])
-    bookmark = current_user.bookmarks.find_by(sauna_id: sauna.id)
-    bookmark.destroy
-    redirect_back(fallback_location: root_path)
+    if user_signed_in?
+      sauna = Sauna.find(params[:sauna_id])
+      bookmark = current_user.bookmarks.find_by(sauna_id: sauna.id)
+      bookmark.destroy
+      redirect_back(fallback_location: root_path)
+    else
+      redirect_to new_user_session_path
+    end
+
   end
 end
